add missing qreal()
on ARM qreal != double so a qbound with qreal,double,qreal will on
arm be a qbound on float,double,float which defeats the exercise
of making them qreal to begin with.

CCMAIL: francesco.cecconi@gmail.com

+++ src/mpris2/mediaplayer2player.cpp
@@ -191,7 +191,7 @@ double MediaPlayer2Player::Volume() const
 
 void MediaPlayer2Player::setVolume(double volume) const
 {
-    Dragon::engine()->setVolume(qBound(qreal(0.0), volume, qreal(1.0)));
+    Dragon::engine()->setVolume(qBound(qreal(0.0), qreal(volume), qreal(1.0)));
 }
 
 qlonglong MediaPlayer2Player::Position() const
